+++ srcs/utils/utils3.c
@@ -6,7 +6,7 @@
 /*   By: wouhliss <wouhliss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/23 18:23:37 by ybelatar          #+#    #+#             */
-/*   Updated: 2024/04/04 02:20:38 by wouhliss         ###   ########.fr       */
+/*   Updated: 2024/04/05 11:40:38 by wouhliss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,9 +49,9 @@ void	*ft_calloc(size_t nmemb, size_t size)
 	return (res);
 }
 
-inline void	ft_swapd(float *a, float *b)
+inline void	ft_swapd(double *a, double *b)
 {
-	float	c;
+	double	c;
 
 	c = *a;
 	*a = *b;
